FIX : Eliminating awaitUninterruptibly() call from the MockServerHandler class, causing BlockingOperationException in some cases to prevent dead locks.
According to the Java doc of Netty ChannelHandlerFuture class:

--------------------------------------- copy ------------------------------------------------
Do not call await() inside ChannelHandler

The event handler methods in ChannelHandler is usually called by an I/O thread. If await() is called by an event handler method, which is called by the I/O thread, the I/O operation it is waiting for might never be complete because await() can block the I/O operation it is waiting for, which is a dead lock.

 // BAD - NEVER DO THIS
 @Override
 public void channelRead(ChannelHandlerContext ctx, GoodByeMessage msg) {
     ChannelFuture future = ctx.channel().close();
     future.awaitUninterruptibly();
     // Perform post-closure operation
     // ...
 }

 // GOOD
 @Override
 public void channelRead(ChannelHandlerContext ctx,  GoodByeMessage msg) {
     ChannelFuture future = ctx.channel().close();
     future.addListener(new ChannelFutureListener() {
         public void operationComplete(ChannelFuture future) {
             // Perform post-closure operation
             // ...
         }
     });
 }
 --------------------------------------- copy ------------------------------------------------

I removed the awaitUninterruptibly() call, because ctx.writeAndFlush(response).addListener(ChannelFutureListener.CLOSE) is enough.