Added the ability to get Expectations back from retrieve for a proxy /retrieve?type=expectations.
This will pull ALL httprequest/httpresponse combinations back for a matching httprequest (or all if httprequest is null).  This is a little different then what dumpToLog was doing, but is going to be used to allow for retrieval of data from a docker container without having to print anything to the logs.

Our use case for this is fairly simple.

As a developer I would like the proxy configured mockserver to provide me every request/response combination I have made to validate an integration test has not changed, thereby allowing me to understand when contracts between services have evolved.