Decrease the object depth when finishing the base array
The parameters is an array of objects, so when we hit end array it
should be the base array.  If this doesn't decrement the depth, the loop
will continue and the next getCurrentToken will take the "type" that is
next in the "body" hash and thus will ignore it until it hits the end of
the body that will finally decrement de object depth.  So everything
after "parameters" was curerntly ignored!

Fixes #371